WIP flakey test
